Avoid (very unlikely) overflow

+++ include/libalx/base/stdlib/alloc/mallocarray.h
@@ -67,7 +67,7 @@ void	*alx_mallocarray	(ptrdiff_t nmemb, size_t size)
 		goto ovf;
 	if (!size)
 		return	NULL;
-	if ((size_t)nmemb > (SIZE_MAX / size))
+	if ((uintmax_t)nmemb > (SIZE_MAX / size))
 		goto ovf;
 
 	return	malloc(size * (size_t)nmemb);
